Report location already used error in a nicer form

+++ actions.js
@@ -481,6 +481,7 @@ function installer(app, configure, manifest, appStoreId, waitForHealthcheck, ins
             return req.send(data);
         }, function (error, result) {
             if (error) exit(error);
+            if (result.statusCode === 409) exit(util.format('Failed to install app. The location %s is already used.'.red, location.bold));
             if (result.statusCode !== 202) exit(util.format('Failed to install app.'.red, result.statusCode, result.text));
 
             var appId = app ? app.id : result.body.id;
